requirements.txt
# Core dependencies
langgraph>=0.0.27
anthropic>=0.8.0
streamlit>=1.30.0
langchain>=0.1.0
langchain-anthropic>=0.1.1

# Document processing
mistralai>=0.0.8
pypdf>=3.17.0
pytesseract>=0.3.10
pillow>=10.0.0

# Vector database for RAG
chromadb>=0.4.18
langchain-chroma>=0.0.1

# Utility libraries
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
tenacity>=8.2.3

# Development tools
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.1
isort>=5.12.0